requirements.txt
django-filter
asgiref==3.2.3
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0
django-crispy-forms==1.7.2
django-import-export==1.2.0
django-widget-tweaks==1.4.5
et-xmlfile==1.0.1
idna==2.8
jdcal==1.4.1
Markdown==3.1.1
MarkupPy==1.14
odfpy==1.4.0
openpyxl==3.0.2
Pillow==6.1.0
pytz==2019.1
PyYAML==5.1.2
requests==2.22.0
sqlparse==0.3.0
tablib==0.14.0
urllib3==1.25.3
virtualenv==16.6.2
xlrd==1.2.0
xlwt==1.3.0
